pyproject.toml
[tool.poetry]
name = "vector-vault"
version = "0.1.0"
description = ""
authors = ["Rami"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.2.3"
pgvector = "0.4.1"
drf-spectacular = "0.28.0"
djangorestframework-simplejwt = "5.5.0"
psycopg2 = "2.9.10"
dj-database-url = "2.3.0"
gunicorn = "23.0.0"
whitenoise = "6.9.0"
django-environ = "0.12.0"
sentence-transformers = "4.1.0"

[tool.poetry.group.dev.dependencies]
django-silk = "^5.3.2"
pre-commit = "^4.1.0"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
